requirements.txt
fastapi
uvicorn
pillow
numpy==1.24.4
tensorflow-cpu==2.14.0
matplotlib
python-multipart